Moving fixture files to specific folders to make ava ignore them
