+++ src/__tests__/fixtures/syntax-error/1_syntax.sql
